+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Provide partial support for Opal Ruby (missing display of file name, line number, and source code)
 - `source_line_count` option
 - `wraper` option for including both `header` and `footer`
+- Special handling of exceptions (prints using full_message)
 
 ## 0.8.2